deleting useless files and optimizing web
