+++ src/assets/translations/en.translations.json
@@ -1,37 +1,17 @@
 {
-  "header.when": "When",
-  "header.where": "Where",
-  "header.how": "How",
+  "header.thanks": "Thanks",
   "header.gallery": "Gallery",
   "confirm.button": "Confirm Attendance",
-  "when.date": "Sunday, october 27, 2024",
-  "when.intro.1": "Two origins, two cultures, two languages, but one destiny.",
-  "when.intro.2": "Spain, halfway between France and Mexico, the rightful heir to our future world.",
-  "when.intro.3": "With joy, we invite you to our wedding ceremony, which will take place at Castell de Tous in Barcelona.",
-  "when.intro.4": "Your presence and support on this very special day mean a lot to us.",
-  "when.details.1": "We are happy to invite you to our wedding ceremony, which will take place at Castell de Tous, in Barcelona.",
-  "when.details.2": "Afterwards, we will celebrate together with a reception starting at 15h.",
-  "when.details.3": "It is not necessary to bring a gift, as we know that everyone is traveling from different parts of the world, and we consider your presence our greatest present.",
-  "when.details.4": "Thank you for joining us!",
-  "when.ceremony.1": "CEREMONY",
-  "when.ceremony.2": "14:00h",
-  "when.ceremony.3": "Castell de Tous",
-  "when.reception.1": "RECEPTION",
-  "when.reception.2": "15:00h",
-  "when.reception.3": "Castell de Tous",
-  "where.text.1": "CASTELL DE TOUS",
-  "where.text.2": "14:00h",
-  "where.text.3": "How to get there",
-  "how.text.1": "The wedding will take place on Sunday, October 27th, at Castell de Tous, Barcelona.",
-  "how.text.2": "A private transportation service will depart from Av. Reina Maria Cristina, directly in front of the Venetian Towers in Barcelona.",
-  "how.text.3": "If you choose to drive, there will be parking available at Castell de Tous.",
-  "how.text.4": "PRIVATE BUS",
-  "how.text.5": "Meeting Time: 11:45h",
-  "how.text.6": "Bus Departure: 12:00h",
-  "how.text.7": "Meeting Point",
-  "how.text.8": "CEREMONY",
-  "how.text.9": "14:00h ",
-  "how.text.10": "Castell de Tous, Barcelone",
+  "thanks.text.1": "The wedding will take place on Sunday, October 27th, at Castell de Tous, Barcelona.",
+  "thanks.text.2": "A private transportation service will depart from Av. Reina Maria Cristina, directly in front of the Venetian Towers in Barcelona.",
+  "thanks.text.3": "If you choose to drive, there will be parking available at Castell de Tous.",
+  "thanks.text.4": "PRIVATE BUS",
+  "thanks.text.5": "Meeting Time: 11:45h",
+  "thanks.text.6": "Bus Departure: 12:00h",
+  "thanks.text.7": "Meeting Point",
+  "thanks.text.8": "CEREMONY",
+  "thanks.text.9": "14:00h ",
+  "thanks.text.10": "Castell de Tous, Barcelone",
   "modal.text.1": "CONFIRM ATTENDANCE",
   "modal.text.2": "Please confirm by August 1st, 2024.",
   "modal.text.3": "First Name",
@@ -45,6 +25,6 @@
   "modal.text.11": "Gluten free",
   "modal.text.12": "Vegan",
   "modal.text.13": "Other",
-  "modal.send.btn": "Send",
-  "accommodations.text": "Accommodations"
+  "modal.send.btn": "Send"
+
 }
