+++ src/assets/translations/es.translations.json
@@ -1,37 +1,17 @@
 {
-  "header.when": "Cuándo",
-  "header.where": "Dónde",
-  "header.how": "Cómo",
+  "header.thanks": "Gracias",
   "header.gallery": "Galería",
   "confirm.button": "Confirmar asistencia",
-  "when.date": "Domingo, 27 de octubre de 2024",
-  "when.intro.1": "Dos orígenes, dos culturas, dos idiomas, un destino.",
-  "when.intro.2": "España, a medio camino entre Francia y México, destinatario legítimo de nuestro mundo por venir.",
-  "when.intro.3": "Con alegría, los invitamos a nuestra ceremonia de boda que tendrá lugar en el Castell de Tous en Barcelona.",
-  "when.intro.4": "Agradecemos profundamente su presencia y apoyo en este día tan especial.",
-  "when.details.1": "Estamos felices de invitarte a nuestra ceremonia de boda, que tendrá lugar en Castell de Tous, en Barcelona.",
-  "when.details.2": "Después celebraremos juntos con una recepción a partir de las 15h.",
-  "when.details.3": "No es necesario traer un regalo, ya que sabemos que todos viajan desde distintas partes del mundo y consideramos tu presencia nuestro mejor obsequio.",
-  "when.details.4": "¡Gracias por acompañarnos!",
-  "when.ceremony.1": "CEREMONIA",
-  "when.ceremony.2": "14:00h",
-  "when.ceremony.3": "Castell de Tous",
-  "when.reception.1": "RECEPCIÓN",
-  "when.reception.2": "15:00h",
-  "when.reception.3": "Castell de Tous",
-  "where.text.1": "CASTELL DE TOUS",
-  "where.text.2": "14:00h",
-  "where.text.3": "Cómo llegar",
-  "how.text.1": "La boda tendrá lugar el domingo 27 de octubre en el Castell de Tous, Barcelona.",
-  "how.text.2": "Un servicio de transporte privado saldrá desde la Av. Reina María Cristina, justo frente a las Torres Venecianas en Barcelona.",
-  "how.text.3": "Si decides conducir, habrá estacionamiento disponible en el Castell de Tous.",
-  "how.text.4": "AUTOBÚS PRIVADO",
-  "how.text.5": "Hora de encuentro: 11:45h",
-  "how.text.6": "Salida del autobús: 12:00h",
-  "how.text.7": "Punto de encuentro",
-  "how.text.8": "CEREMONIA",
-  "how.text.9": "14:00h ",
-  "how.text.10": "Castell de Tous, Barcelona",
+  "thanks.text.1": "La boda tendrá lugar el domingo 27 de octubre en el Castell de Tous, Barcelona.",
+  "thanks.text.2": "Un servicio de transporte privado saldrá desde la Av. Reina María Cristina, justo frente a las Torres Venecianas en Barcelona.",
+  "thanks.text.3": "Si decides conducir, habrá estacionamiento disponible en el Castell de Tous.",
+  "thanks.text.4": "AUTOBÚS PRIVADO",
+  "thanks.text.5": "Hora de encuentro: 11:45h",
+  "thanks.text.6": "Salida del autobús: 12:00h",
+  "thanks.text.7": "Punto de encuentro",
+  "thanks.text.8": "CEREMONIA",
+  "thanks.text.9": "14:00h ",
+  "thanks.text.10": "Castell de Tous, Barcelona",
   "modal.text.1": "CONFIRMAR ASISTENCIA",
   "modal.text.2": "Por favor confirma antes del 1 de agosto de 2024.",
   "modal.text.3": "Nombre",
@@ -45,6 +25,6 @@
   "modal.text.11": "Libre de gluten",
   "modal.text.12": "Vegano",
   "modal.text.13": "Otro",
-  "modal.send.btn": "Enviar",
-  "accommodations.text": "Alojamientos"
+  "modal.send.btn": "Enviar"
+
 }
