+++ src/assets/translations/fr.translations.json
@@ -1,37 +1,17 @@
 {
-  "header.when": "Quand",
-  "header.where": "Où",
-  "header.how": "Comment",
+  "header.thanks": "Merci",
   "header.gallery": "Galerie",
   "confirm.button": "Confirmer la présence",
-  "when.date": "Dimanche, 27 octobre 2024",
-  "when.intro.1": "Deux origines, deux cultures, deux langues, mais un seul destin.",
-  "when.intro.2": "Espagne, à mi-chemin entre la France et le Mexique, destinataire légitime de notre monde en devenir.",
-  "when.intro.3": "C'est avec joie que nous vous invitons à notre cérémonie de mariage qui aura lieu au Castell de Tous à Barcelone.",
-  "when.intro.4": "Nous vous remercions profondément de votre présence et de votre soutien en cette journée si spéciale.",
-  "when.details.1": "Nous sommes heureux de vous inviter à notre cérémonie de mariage, qui aura lieu au Castell de Tous, à Barcelone.",
-  "when.details.2": "Ensuite, nous célébrerons ensemble avec une réception à partir de 15h.",
-  "when.details.3": "Il n'est pas nécessaire d'apporter un cadeau, car nous savons que vous voyagez tous depuis différentes parties du monde et nous considérons votre présence comme notre plus beau cadeau.",
-  "when.details.4": "Merci de nous accompagner !",
-  "when.ceremony.1": "CÉRÉMONIE",
-  "when.ceremony.2": "14:00h",
-  "when.ceremony.3": "Castell de Tous",
-  "when.reception.1": "RÉCEPTION",
-  "when.reception.2": "15:00h",
-  "when.reception.3": "Castell de Tous,",
-  "where.text.1": "CASTELL DE TOUS",
-  "where.text.2": "14:00h",
-  "where.text.3": "Comment y arriver ",
-  "how.text.1": "Le mariage aura lieu le dimanche 27 octobre au Castell de Tous, à Barcelone.",
-  "how.text.2": "Un service de transport privé partira de l'Av. Reina Maria Cristina, juste en face des Tours Vénitiennes à Barcelone.",
-  "how.text.3": "Si vous choisissez de conduire, un parking sera disponible au Castell de Tous.",
-  "how.text.4": "BUS PRIVÉ",
-  "how.text.5": "Heure de rendez-vous : 11h45",
-  "how.text.6": "Départ du bus : 12h00",
-  "how.text.7": "Point de rendez-vous",
-  "how.text.8": "CÉRÉMONIE",
-  "how.text.9": "14:00h",
-  "how.text.10": "Castell de Tous, Barcelone",
+  "thanks.text.1": "Le mariage aura lieu le dimanche 27 octobre au Castell de Tous, à Barcelone.",
+  "thanks.text.2": "Un service de transport privé partira de l'Av. Reina Maria Cristina, juste en face des Tours Vénitiennes à Barcelone.",
+  "thanks.text.3": "Si vous choisissez de conduire, un parking sera disponible au Castell de Tous.",
+  "thanks.text.4": "BUS PRIVÉ",
+  "thanks.text.5": "Heure de rendez-vous : 11h45",
+  "thanks.text.6": "Départ du bus : 12h00",
+  "thanks.text.7": "Point de rendez-vous",
+  "thanks.text.8": "CÉRÉMONIE",
+  "thanks.text.9": "14:00h",
+  "thanks.text.10": "Castell de Tous, Barcelone",
   "modal.text.1": "CONFIRMER LA PRÉSENCE",
   "modal.text.2": "Veuillez confirmer d'ici le 1er août 2024.",
   "modal.text.3": "Prénom",
@@ -45,6 +25,6 @@
   "modal.text.11": "Sans gluten",
   "modal.text.12": "Végétalien",
   "modal.text.13": "Autre",
-  "modal.send.btn": "Envoyer",
-  "accommodations.text": "Hébergements"
+  "modal.send.btn": "Envoyer"
+
 }
